requirements.txt
jinja2>=3.1
pyyaml>=6.0
pillow>=10.0      # for photo preprocessing (contrast boost, auto-crop)
